requirements.txt
caldav==1.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
icalendar==5.0.10
idna==3.4
install==1.3.5
lxml==4.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
recurring-ical-events==2.1.0
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.0.7
vobject==0.9.6.1
x-wr-timezone==0.0.5
